+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 2.2.0 (in progress)
 - Add Probability module.
+- Add Gate Delay module (in progress).
 - Fix polyphony for individual outs on Vector Mix.
+- Add missing polyphony tag to Vector Mix.
 
 ## 2.1.0
 - Add Apple Silicon support to build script.